pyproject.toml
[project]
name = "nasa-slater-catalog"
version = "0.1.0"
description = "Cataloging pipeline for NASA archival video using local GPU inference"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "marker-pdf>=1.7,<1.11",
    "openpyxl>=3.1.5",
    "torch>=2.5",
    "torchvision>=0.20",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu126", marker = "sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu126", marker = "sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
